Unbreak Win64 build.
Review URL: https://codereview.chromium.org/13976013

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@14300 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/runtime.cc
@@ -825,7 +825,8 @@ RUNTIME_FUNCTION(MaybeObject*, Runtime_TypedArrayInitialize) {
   ASSERT(byte_length % elementSize == 0);
   size_t length = byte_length / elementSize;
 
-  Handle<Object> length_obj = isolate->factory()->NewNumber(length);
+  Handle<Object> length_obj =
+      isolate->factory()->NewNumber(static_cast<double>(length));
   holder->set_length(*length_obj);
   Handle<ExternalArray> elements =
       isolate->factory()->NewExternalArray(
